Add repeat option
In particular cases (such as a very short count) the `--pv-update-period` option might be not enough to have a PV value in the NeXus file.
This change introduce a feature that forces the forwarder to repeat all the Kafka Logs currently cached. This is triggered with the following JSON message in the config-topic:
{"command_type": "repeat"}