CI: Fix mypy numpy types
Numpy offers a numpy plugin numpy.typing.mypy_plugin
We load it in pyproject.toml but it seems to not work when the plugin
name has quotes. Without quotes, the TOML file is invalid for tomllib
and breaks black and flake8.

We therefore add the plugin with quotes, hoping that mypy will eventually
use it correctly.
We also add a few numpy types manually to silence mypy warning.